requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
apipkg==1.4               # via execnet
asn1crypto==0.24.0        # via cryptography
attrs==17.4.0             # via pytest
bleach==1.5.0             # via wiki
cffi==1.11.5              # via cryptography
click==6.7                # via pip-tools
coverage==4.3.1
cryptography==2.1.4       # via pyopenssl
decorator==4.2.1          # via ipython, traitlets
django-autocomplete-light==3.2.1
django-ckeditor==5.4.0
django-classy-tags==0.8.0  # via django-sekizai
django-crontab==0.7.1
django-debug-toolbar==1.6
django-extended-choices==1.1.1
django-extensions==2.0.0
django-hashers-passlib==0.3
django-js-asset==1.0.0    # via django-ckeditor
django-markdown2==0.3.0
django-material==1.1.0
django-model-utils==3.0.0
django-mptt==0.8.7        # via wiki
django-nyt==1.0           # via wiki
django-post-office==2.0.8
django-recaptcha==1.2.0
django-sekizai==0.10.0    # via wiki
django-smart-selects==1.3.2
django==1.11
djangorestframework==3.7.7
execnet==1.5.0            # via pytest-xdist
factory-boy==2.9.2
faker==0.8.11             # via factory-boy
first==2.0.1              # via pip-tools
freezegun==0.3.9
future==0.16.0            # via django-extended-choices
html5lib==0.9999999       # via bleach
idna==2.6                 # via cryptography
ipython-genutils==0.2.0   # via traitlets
ipython==6.2.1
jedi==0.11.1              # via ipython
jsonfield==2.0.2          # via django-post-office
markdown2==2.3.5          # via django-markdown2
markdown==2.6.11
olefile==0.44
parso==0.1.1              # via jedi
passlib==1.7.1            # via django-hashers-passlib
pexpect==4.4.0            # via ipython
pickleshare==0.7.4        # via ipython
pillow==5.0.0             # via wiki
pip-tools==1.10.1
pluggy==0.6.0             # via pytest
prompt-toolkit==1.0.15    # via ipython
psycopg2==2.6.2
ptyprocess==0.5.2         # via pexpect
py==1.5.2                 # via pytest
pycparser==2.18           # via cffi
pygments==2.2.0           # via ipython
pyopenssl==17.5.0
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-forked==0.2        # via pytest-xdist
pytest-xdist==1.20.1
pytest==3.3.0
python-dateutil==2.6.1    # via faker, freezegun
pytz==2018.3              # via django
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via bleach, cryptography, django-extensions, django-nyt, faker, freezegun, html5lib, pip-tools, prompt-toolkit, pyopenssl, pytest, python-dateutil, traitlets, wiki
sorl-thumbnail==12.4.1    # via wiki
sqlparse==0.2.4           # via django-debug-toolbar
text-unidecode==1.2       # via faker
traitlets==4.3.2          # via ipython
typing==3.6.4             # via django-extensions
wcwidth==0.1.7            # via prompt-toolkit
werkzeug==0.13
wiki==0.3
